+++ main.py
@@ -1,2 +1,8 @@
-def working():
-    return "working..."
+from game.interfaz import Interfaz
+
+def main():
+    interfaz = Interfaz()
+    interfaz.menu()
+
+if __name__ == "__main__":
+    main()